requirements.txt
Django==4.2.7
djangorestframework==3.15.1
requests==2.31.0
djangorestframework-simplejwt==5.3.1